Refactor canvas and add export utilities
