+++ pages/canvas.js
@@ -1,5 +1,6 @@
 import React, { useRef, useState, useEffect } from 'react';
-import { jsPDF } from 'jspdf';
+import { savePDF, exportHTML } from '../utils/export';
+import { pointToShape, bounds, cornerHit, getCornerPos, drawShape } from '../utils/shapes';
 import { TbRectangle, TbSquare, TbCircle, TbTriangle, TbDiamond, TbPentagon, TbHexagon, TbOctagon, TbStar, TbArrowRight, TbHeart, TbTypography } from 'react-icons/tb';
 import { IoEllipse } from 'react-icons/io5';
 import { BsHeptagon } from 'react-icons/bs';
@@ -47,59 +48,22 @@ export default function CanvasPage() {
   const [shapeStart, setShapeStart] = useState(null);
   const [rotateId, setRotateId] = useState(null);
   const [rotateStart, setRotateStart] = useState(null);
+  const [previewShape, setPreviewShape] = useState(null);
 
-  const pointToShape = (shape, px, py) => {
-    const angle = (-shape.rotation * Math.PI) / 180;
-    const dx = px - shape.x;
-    const dy = py - shape.y;
-    return {
-      x: dx * Math.cos(angle) - dy * Math.sin(angle),
-      y: dx * Math.sin(angle) + dy * Math.cos(angle),
-    };
-  };
-
-  const bounds = (shape) => {
-    const w = shape.width || shape.radius * 2 || shape.fontSize * (shape.text?.length || 1);
-    const h = shape.height || shape.radius * 2 || shape.fontSize;
-    return { w, h };
-  };
-
-  const cornerHit = (shape, px, py) => {
-    const { w, h } = bounds(shape);
-    const p = pointToShape(shape, px, py);
-    const size = 6;
-    if (Math.abs(p.x + w / 2) <= size && Math.abs(p.y + h / 2) <= size) return 'nw';
-    if (Math.abs(p.x - w / 2) <= size && Math.abs(p.y + h / 2) <= size) return 'ne';
-    if (Math.abs(p.x - w / 2) <= size && Math.abs(p.y - h / 2) <= size) return 'se';
-    if (Math.abs(p.x + w / 2) <= size && Math.abs(p.y - h / 2) <= size) return 'sw';
-    if (Math.abs(p.x) <= size && Math.abs(p.y + h / 2 + 20) <= size) return 'rotate';
-    return null;
-  };
-
-  const getCornerPos = (shape, corner) => {
-    const { w, h } = bounds(shape);
-    const map = {
-      nw: { x: -w / 2, y: -h / 2 },
-      ne: { x: w / 2, y: -h / 2 },
-      se: { x: w / 2, y: h / 2 },
-      sw: { x: -w / 2, y: h / 2 },
-    };
-    const pt = map[corner];
-    const angle = (shape.rotation * Math.PI) / 180;
-    return {
-      x: shape.x + pt.x * Math.cos(angle) - pt.y * Math.sin(angle),
-      y: shape.y + pt.x * Math.sin(angle) + pt.y * Math.cos(angle),
-    };
-  };
 
   useEffect(() => {
     const canvas = canvasRef.current;
     const ctx = canvas.getContext('2d');
     ctx.clearRect(0, 0, canvas.width, canvas.height);
     shapes.forEach((shape) => {
-      drawShape(ctx, shape);
+      drawShape(ctx, shape, selectedId);
     });
-  }, [shapes]);
+    if (previewShape) {
+      ctx.globalAlpha = 0.5;
+      drawShape(ctx, previewShape);
+      ctx.globalAlpha = 1;
+    }
+  }, [shapes, previewShape, selectedId]);
 
   useEffect(() => {
     const canvas = canvasRef.current;
@@ -159,8 +123,47 @@ export default function CanvasPage() {
       setSelectedId(null);
     };
     const handleMove = (e) => {
-      if (drawingImage || (drawingShape && shapeStart)) return;
       const { x, y } = getPos(e);
+      if (drawingImage && imageStart && pendingImage) {
+        const preview = {
+          id: 'preview',
+          type: 'image',
+          x: (imageStart.x + x) / 2,
+          y: (imageStart.y + y) / 2,
+          width: Math.abs(x - imageStart.x),
+          height: Math.abs(y - imageStart.y),
+          src: pendingImage.src,
+          img: pendingImage.img,
+          rotation: 0,
+          opacity: 1,
+          strokeWidth: 0,
+          lineDash: 0,
+        };
+        setPreviewShape(preview);
+        return;
+      }
+      if (drawingShape && shapeStart && pendingShape) {
+        const w = Math.abs(x - shapeStart.x);
+        const h = Math.abs(y - shapeStart.y);
+        const midx = (x + shapeStart.x) / 2;
+        const midy = (y + shapeStart.y) / 2;
+        const preview = { ...pendingShape, x: midx, y: midy };
+        if (['circle','pentagon','hexagon','heptagon','octagon','star'].includes(pendingShape.type)) {
+          preview.radius = Math.max(w, h) / 2;
+        } else if (pendingShape.type === 'square') {
+          const side = Math.max(w, h);
+          preview.width = side;
+          preview.height = side;
+        } else if (pendingShape.type === 'text') {
+          preview.fontSize = Math.max(5, Math.max(w, h));
+        } else {
+          preview.width = w;
+          preview.height = h;
+        }
+        setPreviewShape(preview);
+        return;
+      }
+      setPreviewShape(null);
       if (rotateId !== null && rotateStart) {
         const sel = shapes.find((s) => s.id === rotateId);
         if (sel) {
@@ -231,6 +234,7 @@ export default function CanvasPage() {
         setDrawingImage(false);
         setImageStart(null);
         setPendingImage(null);
+        setPreviewShape(null);
         return;
       }
       if (drawingShape && shapeStart && pendingShape) {
@@ -256,11 +260,13 @@ export default function CanvasPage() {
         setDrawingShape(false);
         setShapeStart(null);
         setPendingShape(null);
+        setPreviewShape(null);
         return;
       }
       setDraggingId(null);
       setResizingId(null);
       setRotateId(null);
+      setPreviewShape(null);
     };
     canvas.addEventListener('mousedown', handleDown);
     canvas.addEventListener('mousemove', handleMove);
@@ -272,173 +278,6 @@ export default function CanvasPage() {
     };
   }, [shapes, draggingId, dragOffset, drawingImage, imageStart, pendingImage, resizingId, resizeStart, drawingShape, shapeStart, pendingShape, selectedId, rotateId, rotateStart]);
 
-  const drawShape = (ctx, shape) => {
-    ctx.save();
-    ctx.translate(shape.x, shape.y);
-    ctx.rotate((shape.rotation * Math.PI) / 180);
-    ctx.globalAlpha = shape.opacity;
-    ctx.lineWidth = shape.strokeWidth;
-    ctx.strokeStyle = shape.strokeColor;
-    ctx.fillStyle = shape.fillColor;
-    if (shape.lineDash > 0) {
-      ctx.setLineDash([shape.lineDash]);
-    } else {
-      ctx.setLineDash([]);
-    }
-    switch (shape.type) {
-      case 'rectangle':
-        ctx.fillRect(-shape.width/2, -shape.height/2, shape.width, shape.height);
-        ctx.strokeRect(-shape.width/2, -shape.height/2, shape.width, shape.height);
-        break;
-      case 'square':
-        ctx.fillRect(-shape.width/2, -shape.width/2, shape.width, shape.width);
-        ctx.strokeRect(-shape.width/2, -shape.width/2, shape.width, shape.width);
-        break;
-      case 'circle':
-        ctx.beginPath();
-        ctx.arc(0, 0, shape.radius, 0, 2 * Math.PI);
-        ctx.fill();
-        ctx.stroke();
-        break;
-      case 'ellipse':
-        ctx.beginPath();
-        ctx.ellipse(0, 0, shape.width/2, shape.height/2, 0, 0, 2 * Math.PI);
-        ctx.fill();
-        ctx.stroke();
-        break;
-      case 'triangle':
-        ctx.beginPath();
-        ctx.moveTo(0, -shape.height/2);
-        ctx.lineTo(shape.width/2, shape.height/2);
-        ctx.lineTo(-shape.width/2, shape.height/2);
-        ctx.closePath();
-        ctx.fill();
-        ctx.stroke();
-        break;
-      case 'diamond':
-        ctx.beginPath();
-        ctx.moveTo(0, -shape.height/2);
-        ctx.lineTo(shape.width/2, 0);
-        ctx.lineTo(0, shape.height/2);
-        ctx.lineTo(-shape.width/2, 0);
-        ctx.closePath();
-        ctx.fill();
-        ctx.stroke();
-        break;
-      case 'pentagon':
-      case 'hexagon':
-      case 'heptagon':
-      case 'octagon':
-        const sidesMap = { pentagon:5, hexagon:6, heptagon:7, octagon:8 };
-        const sides = sidesMap[shape.type];
-        ctx.beginPath();
-        for (let i = 0; i < sides; i++) {
-          const angle = (i * 2 * Math.PI) / sides;
-          const px = shape.radius * Math.cos(angle);
-          const py = shape.radius * Math.sin(angle);
-          if (i === 0) ctx.moveTo(px, py); else ctx.lineTo(px, py);
-        }
-        ctx.closePath();
-        ctx.fill();
-        ctx.stroke();
-        break;
-      case 'star':
-        ctx.beginPath();
-        const spikes = 5;
-        const outerRadius = shape.radius;
-        const innerRadius = shape.radius/2;
-        for (let i = 0; i < spikes * 2; i++) {
-          const r = i % 2 === 0 ? outerRadius : innerRadius;
-          const angle = (i * Math.PI) / spikes;
-          const px = r * Math.cos(angle);
-          const py = r * Math.sin(angle);
-          if (i === 0) ctx.moveTo(px, py); else ctx.lineTo(px, py);
-        }
-        ctx.closePath();
-        ctx.fill();
-        ctx.stroke();
-        break;
-      case 'trapezoid':
-        ctx.beginPath();
-        ctx.moveTo(-shape.width/2, -shape.height/2);
-        ctx.lineTo(shape.width/2, -shape.height/2);
-        ctx.lineTo(shape.width/4, shape.height/2);
-        ctx.lineTo(-shape.width/4, shape.height/2);
-        ctx.closePath();
-        ctx.fill();
-        ctx.stroke();
-        break;
-      case 'parallelogram':
-        ctx.beginPath();
-        ctx.moveTo(-shape.width/2 + shape.width/4, -shape.height/2);
-        ctx.lineTo(shape.width/2 + shape.width/4, -shape.height/2);
-        ctx.lineTo(shape.width/2 - shape.width/4, shape.height/2);
-        ctx.lineTo(-shape.width/2 - shape.width/4, shape.height/2);
-        ctx.closePath();
-        ctx.fill();
-        ctx.stroke();
-        break;
-      case 'arrow':
-        ctx.beginPath();
-        ctx.moveTo(-shape.width/2, -shape.height/4);
-        ctx.lineTo(0, -shape.height/4);
-        ctx.lineTo(0, -shape.height/2);
-        ctx.lineTo(shape.width/2, 0);
-        ctx.lineTo(0, shape.height/2);
-        ctx.lineTo(0, shape.height/4);
-        ctx.lineTo(-shape.width/2, shape.height/4);
-        ctx.closePath();
-        ctx.fill();
-        ctx.stroke();
-        break;
-      case 'heart':
-        ctx.beginPath();
-        ctx.moveTo(0, shape.height/4);
-        ctx.bezierCurveTo(shape.width/2, -shape.height/4, shape.width/2, shape.height/2, 0, shape.height/2);
-        ctx.bezierCurveTo(-shape.width/2, shape.height/2, -shape.width/2, -shape.height/4, 0, shape.height/4);
-        ctx.fill();
-        ctx.stroke();
-        break;
-      case 'image':
-        if (shape.img) {
-          ctx.drawImage(shape.img, -shape.width/2, -shape.height/2, shape.width, shape.height);
-        }
-        break;
-      case 'text':
-        ctx.fillStyle = shape.fontColor;
-        ctx.font = `${shape.fontSize}px sans-serif`;
-        ctx.textAlign = 'center';
-        ctx.textBaseline = 'middle';
-        ctx.fillText(shape.text, 0, 0);
-        break;
-      default:
-        break;
-    }
-    if (shape.id === selectedId) {
-      const { w, h } = bounds(shape);
-      const size = 6;
-      ctx.strokeStyle = 'blue';
-      ctx.setLineDash([4]);
-      ctx.strokeRect(-w / 2, -h / 2, w, h);
-      ctx.setLineDash([]);
-      const corners = [
-        [-w / 2, -h / 2],
-        [w / 2, -h / 2],
-        [w / 2, h / 2],
-        [-w / 2, h / 2],
-      ];
-      ctx.fillStyle = 'white';
-      corners.forEach(([cx, cy]) => {
-        ctx.fillRect(cx - size / 2, cy - size / 2, size, size);
-        ctx.strokeRect(cx - size / 2, cy - size / 2, size, size);
-      });
-      ctx.beginPath();
-      ctx.arc(0, -h / 2 - 20, size / 2, 0, 2 * Math.PI);
-      ctx.fill();
-      ctx.stroke();
-    }
-    ctx.restore();
-  };
 
   const addShape = () => {
     setPendingShape({ ...current });
@@ -476,19 +315,14 @@ export default function CanvasPage() {
     URL.revokeObjectURL(url);
   };
 
-  const savePDF = () => {
-    const canvas = canvasRef.current;
-    if (!canvas) return;
-    const orientation = canvas.width > canvas.height ? 'landscape' : 'portrait';
-    const pdf = new jsPDF({
-      orientation,
-      unit: 'px',
-      format: [canvas.width, canvas.height],
-    });
-    const imgData = canvas.toDataURL('image/png');
-    pdf.addImage(imgData, 'PNG', 0, 0, canvas.width, canvas.height);
-    pdf.save('design.pdf');
+  const handleSavePDF = () => {
+    savePDF(canvasRef.current);
   };
+
+  const handleExportHTML = () => {
+    exportHTML(canvasRef.current);
+  };
+
 
   const updateCurrent = (field, value) => {
     setCurrent({ ...current, [field]: value });
@@ -634,7 +468,8 @@ export default function CanvasPage() {
         <button onClick={addShape}>Agregar</button>
         <hr />
         <button onClick={saveJSON}>Guardar JSON</button>
-        <button onClick={savePDF}>Guardar PDF</button>
+        <button onClick={handleSavePDF}>Guardar PDF</button>
+        <button onClick={handleExportHTML}>Exportar HTML</button>
         <input type="file" accept="application/json" onChange={loadJSON} />
       </div>
       <canvas ref={canvasRef} width={800} height={600} style={{ border: '1px solid black', margin: '10px' }} />
